backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
torchvision
transformers
peft
accelerate
